triangles: big UniformBufferObject refactoring
- moved generic buffer stuff to Buffers
- moved descriptor stuff to Descriptor
- moved UniformBufferObject stuff to TransformationObject
- actual TransformationObject type is now implementation detail of
  the module with the same name
- made UBO specific fields of RenderData more generic
- made Lib oblivious of dealing with UBOs
