+++ vulkan-triangles/src/Lib/Vulkan/VertexBuffer.hs
@@ -3,13 +3,10 @@
 {-# LANGUAGE PolyKinds        #-}
 {-# LANGUAGE Strict           #-}
 {-# LANGUAGE TypeApplications #-}
-
 module Lib.Vulkan.VertexBuffer
   ( createVertexBuffer
   , createIndexBuffer
-  , createUniformBuffers
   ) where
-
 
 import           Data.Bits
 import           Foreign.Ptr                              (castPtr)
@@ -87,17 +84,4 @@ createIndexBuffer pdev dev cmdPool cmdQueue (XFrame indices) = do
       copyBuffer dev cmdPool cmdQueue stagingBuf vertexBuf bSize
 
     return vertexBuf
-
-
-createUniformBuffers
-  :: VkPhysicalDevice
-  -> VkDevice
-  -> VkDeviceSize
-  -> Int
-  -> Program r [(VkDeviceMemory, VkBuffer)]
-createUniformBuffers pdev dev bSize n = do
-      sequence $ replicate n $ createBuffer pdev dev bSize
-         VK_BUFFER_USAGE_UNIFORM_BUFFER_BIT
-         ( VK_MEMORY_PROPERTY_HOST_VISIBLE_BIT .|. VK_MEMORY_PROPERTY_HOST_COHERENT_BIT )
-
