+++ vulkan-triangles/vulkan-triangles.cabal
@@ -27,6 +27,8 @@ library
                         Lib.GLFW
                         Lib.Program
                         Lib.Program.Foreign
+                        Lib.Vulkan.Buffer
+                        Lib.Vulkan.Descriptor
                         Lib.Vulkan.Device
                         Lib.Vulkan.Drawing
                         Lib.Vulkan.Instance
@@ -36,6 +38,7 @@ library
                         Lib.Vulkan.Shader.TH
                         Lib.Vulkan.Vertex
                         Lib.Vulkan.VertexBuffer
+                        Lib.Vulkan.TransformationObject
     build-depends:      base >= 4.9 && < 5
                       , vulkan-api >= 1.1.3.0
                       , GLFW-b >= 3.2.1
